style: Rename utils.py as files.py
